+++ Classes/Private/TGPrivateFunctions.m
@@ -8,6 +8,8 @@
 
 #import "TGPrivateFunctions.h"
 #import "TGRESTResource.h"
+#include <sys/sysctl.h>
+#import <mach/mach_time.h>
 
 NSString *TGApplicationDataDirectory(void)
 {
@@ -164,4 +166,35 @@ NSString *TGDestroyRegex(TGRESTResource *resource)
     [regex appendString:[NSString stringWithFormat:@"^(/%@/\\w+/?$)", resource.name]];
     
     return [NSString stringWithString:regex];
+}
+
+uint8_t TGCountOfCores(void)
+{
+    NSUInteger ncpu;
+    size_t len = sizeof(ncpu);
+    sysctlbyname("hw.ncpu", &ncpu, &len, NULL, 0);
+    
+    return ncpu;
+}
+
+CGFloat TGTimedBlock (void (^block)(void))
+{
+    mach_timebase_info_data_t info;
+    if (mach_timebase_info(&info) != KERN_SUCCESS) return -1.0;
+    
+    uint64_t start = mach_absolute_time ();
+    block ();
+    uint64_t end = mach_absolute_time ();
+    uint64_t elapsed = end - start;
+    
+    uint64_t nanos = elapsed * info.numer / info.denom;
+    return (CGFloat)nanos / NSEC_PER_SEC;
+}
+
+CGFloat TGRandomInRange(CGFloat lowerRange, CGFloat upperRange)
+{
+    CGFloat random = drand48();
+    CGFloat range = upperRange - lowerRange;
+    CGFloat scaled = random * range;
+    return scaled + lowerRange;
 }