+++ Tests/TGServerAPIErrorHandlingTests.m
@@ -26,6 +26,8 @@
     [super tearDown];
 }
 
+/*
+
 - (void)testMissingCreateParameters
 {
     
@@ -35,5 +37,7 @@
 {
     
 }
+ 
+ */
 
 @end